requirements.txt
pytest==8.3.4
PyYAML==6.0.2
pytest-testdox==3.1.0
playwright==1.49.1



